cowboy/elixir: Separating The Seas.
