+++ src/elixir/lib/dnsresolvd.ex
@@ -12,43 +12,6 @@
 # (See the LICENSE file at the top of the source tree.)
 #
 
-defmodule AUX do
-    @moduledoc "The helper module for the daemon."
-
-    # Helper constants.
-    def _EXIT_FAILURE    , do:    1
-    def _EXIT_SUCCESS    , do:    0
-    def _EMPTY_STRING    , do:   ""
-    def _COLON_SPACE_SEP , do: ": "
-    def _COMMA_SPACE_SEP , do: ", "
-    def _NEW_LINE        , do: "\n"
-    def _C_FMT           , do: "%c"
-    def _AMPER           , do:  "&"
-    def _ONE_SPACE_STRING, do:  " "
-    def _PRINT_BANNER_OPT, do: "-V"
-
-    # Daemon name, version, and copyright banners.
-    def _DMN_NAME       , do: "DNS Resolver Daemon (dnsresolvd)"
-    def _DMN_DESCRIPTION, do: "Performs DNS lookups for the given "
-                          <>  "hostname passed in an HTTP request"
-    def _DMN_VERSION_S__, do: "Version"
-    def _DMN_VERSION    , do: "0.1"
-    def _DMN_COPYRIGHT__, do: "Copyright (C) 2017-2018"
-    def _DMN_AUTHOR     , do: "Radislav Golubtsov <ragolubtsov@my.com>"
-
-    # Constant: The default hostname to look up for.
-    def _DEF_HOSTNAME, do: "openbsd.org"
-
-    # Helper function. Draws a horizontal separator banner.
-    def _separator_draw(banner_text) do
-        i = String.length(banner_text)
-
-        for (_ <- i..1), do: IO.write('=')
-
-        IO.puts(_EMPTY_STRING())
-    end
-end
-
 defmodule Dnsresolvd do
     @moduledoc "The main module of the daemon."
